+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 * dump with json mode now works.
 * help and text markup is now disabled, which prevents rendering issues when `rich_argparse` is versus isn't installed.
+* Fix issue with type "smartcast:v1" not being respected
 
 
 ## Version 0.8.2 - Released 2025-03-06